+++ modules/gdscript/gdscript_parser.cpp
@@ -4960,8 +4960,10 @@ void GDScriptParser::_parse_class(ClassNode *p_class) {
 						OperatorNode *op = static_cast<OperatorNode *>(subexpr);
 						if (op->op == OperatorNode::OP_CALL && op->arguments[0]->type == Node::TYPE_SELF && op->arguments[1]->type == Node::TYPE_IDENTIFIER) {
 							IdentifierNode *id = static_cast<IdentifierNode *>(op->arguments[1]);
-							if (id->name == "get_node") {
-								_set_error("Use \"onready var " + String(member.identifier) + " = get_node(...)\" instead.");
+							// GOBLIN ENGINE onready warning
+							if (id->name == "get_node" || id->name == "get_child" || id->name == "get_parent") {
+								_set_error("Use \"onready var " + String(member.identifier) + " = " + id->name + "(...)\" instead.");
+								error_line = op->line;
 								return;
 							}
 						}